Move all aws providers to aws provider directory
