+++ resources/aws_providers/aws.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package ctxProvider
+package aws_providers
 
 import "github.com/aws/aws-sdk-go-v2/aws"
