+++ resources/aws_providers/aws_identity_provider.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package ctxProvider
+package aws_providers
 
 import (
 	"context"
