+++ resources/aws_providers/eks_provider.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package fetchers
+package aws_providers
 
 import (
 	"context"
