+++ resources/fetchers/ecr_fetcher.go
@@ -20,6 +20,7 @@ package fetchers
 import (
 	"context"
 	"fmt"
+	"github.com/elastic/cloudbeat/resources/aws_providers"
 	"regexp"
 
 	"github.com/aws/aws-sdk-go-v2/service/ecr"
@@ -34,7 +35,7 @@ const PublicRepoRegex = "public\\.ecr\\.aws\\/\\w+\\/([\\w-]+)\\:?"
 
 type ECRFetcher struct {
 	cfg               ECRFetcherConfig
-	ecrProvider       *ECRProvider
+	ecrProvider       *aws_providers.ECRProvider
 	kubeClient        k8s.Interface
 	repoRegexMatchers []*regexp.Regexp
 }
