+++ resources/fetchers/eks_factory.go
@@ -2,7 +2,7 @@ package fetchers
 
 import (
 	"encoding/gob"
-	"github.com/elastic/cloudbeat/resources/ctxProvider"
+	"github.com/elastic/cloudbeat/resources/aws_providers"
 
 	"github.com/elastic/beats/v7/libbeat/common"
 	"github.com/elastic/cloudbeat/resources/fetching"
@@ -32,9 +32,9 @@ func (f *EKSFactory) Create(c *common.Config) (fetching.Fetcher, error) {
 }
 
 func (f *EKSFactory) CreateFrom(cfg EKSFetcherConfig) (fetching.Fetcher, error) {
-	awsCredProvider := ctxProvider.AWSCredProvider{}
+	awsCredProvider := aws_providers.AWSCredProvider{}
 	awsCfg := awsCredProvider.GetAwsCredentials()
-	eks := NewEksProvider(awsCfg.Config)
+	eks := aws_providers.NewEksProvider(awsCfg.Config)
 
 	fe := &EKSFetcher{
 		cfg:         cfg,
